Update randomSeed.adoc
I like that the original was quite simple, and on its face, seems like a decent way of getting a sort-of random seed.

However, reality sets in when you realize that background noise won't be very likely to vary across the range of the input, even with aref and the analog pin floating.  On my due, I got values clustered around 250.  Worse, this example code is propagated across the internet, leading one to not question how well it works.

The updated code is somewhat more complicated (But still pretty simple), but on a small number (under 100) successive runs, I have not had a repeated value.  

I still wouldn't call this seed "random" in the cryptographically secure context, but I do think its pretty good for seeding an LCRNG.  Probably even better than the typical Desktop C example of srand(time(NULL)); 

It just drove me crazy enough that this example was everywhere and didn't "work" (give me different sequences), that I figured I'd try my first documentation update.  I get that arduino's are for all skill levels, so we try to keep examples as simple as possible, so if that's the case, can we add this code to an "advanced usage" section with some verbiage explaining when to use this version vs the simple version?  I was ready to just input time(NULL) from my PC over the serial line before I realized there was a simple software solution!